+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 更新日志 / Change Log
 
 1. 修复已知BUG / Fix known bugs
-2. 修复无法保存养号设置的问题 / Fix the problem of not being able to save the train settings of the group
+2. 修复设置头像时的BUG / Fix the bug when setting the avatar
+3. 增加清空相册快捷指令 / Add clear gallery shortcut command